rename component file for top page
